Fixed in migrations

+++ src/database/migrations/2015_03_13_122459_create_page_urls_table.php
@@ -14,7 +14,7 @@ class CreatePageUrlsTable extends Migration
     {
         Schema::create('page_urls', function (Blueprint $table) {
             $table->increments('id');
-            $table->smallInteger('page_id')->unsigned();
+            $table->integer('page_id')->unsigned();
             $table->string('location', 2048)->nullable()->index('page_uri_uri');
             $table->boolean('is_primary')->nullable();
             $table->index(['page_id','is_primary'], 'page_uri_page_id_primary_uri');
