+++ src/database/migrations/2015_06_01_210200_add_parent_id_to_pages_table.php
@@ -15,7 +15,7 @@ class AddParentIdToPagesTable extends Migration
     public function up()
     {
         Schema::table('pages', function (Blueprint $table) {
-            $table->smallInteger('parent_id')->unsigned()->nullable();
+            $table->integer('parent_id')->unsigned()->nullable();
             $table->foreign('parent_id')->references('id')->on('pages')->onUpdate('CASCADE')->onDelete('CASCADE');
         });
